+++ src/main/java/net/geforcemods/securitycraft/blockentities/DisguisableBlockEntity.java
@@ -20,6 +20,7 @@ import net.minecraft.world.level.block.state.BlockState;
 import net.minecraft.world.level.block.state.properties.BlockStateProperties;
 import net.minecraft.world.level.material.Fluids;
 import net.neoforged.neoforge.client.model.data.ModelData;
+import net.neoforged.neoforge.common.world.AuxiliaryLightManager;
 import net.neoforged.neoforge.network.PacketDistributor;
 
 public class DisguisableBlockEntity extends CustomizableBlockEntity {
@@ -43,6 +44,7 @@ public class DisguisableBlockEntity extends CustomizableBlockEntity {
 		BlockState state = be.getBlockState();
 		Level level = be.getLevel();
 		BlockPos worldPosition = be.getBlockPos();
+		int newLight = DisguisableBlock.getDisguisedBlockStateFromStack(level, stack).map(s -> s.getLightEmission(level, worldPosition)).orElse(0);
 
 		if (!level.isClientSide) {
 			PacketDistributor.TRACKING_CHUNK.with(level.getChunkAt(worldPosition)).send(new RefreshDisguisableModel(worldPosition, true, stack, toggled));
@@ -52,12 +54,11 @@ public class DisguisableBlockEntity extends CustomizableBlockEntity {
 				level.updateNeighborsAt(worldPosition, state.getBlock());
 			}
 		}
-		else {
+		else
 			ClientHandler.putDisguisedBeRenderer(be, stack);
 
-			if (state.getLightEmission(level, worldPosition) > 0)
-				level.getChunkSource().getLightEngine().checkBlock(worldPosition);
-		}
+		if (newLight > 0)
+			level.getAuxLightManager(worldPosition).setLightAt(worldPosition, newLight);
 	}
 
 	@Override
@@ -69,12 +70,11 @@ public class DisguisableBlockEntity extends CustomizableBlockEntity {
 	}
 
 	public static void onDisguiseModuleRemoved(BlockEntity be, ItemStack stack, boolean toggled) {
+		BlockState state = be.getBlockState();
 		Level level = be.getLevel();
 		BlockPos worldPosition = be.getBlockPos();
 
 		if (!level.isClientSide) {
-			BlockState state = be.getBlockState();
-
 			PacketDistributor.TRACKING_CHUNK.with(level.getChunkAt(worldPosition)).send(new RefreshDisguisableModel(worldPosition, false, stack, toggled));
 
 			if (state.hasProperty(BlockStateProperties.WATERLOGGED) && state.getValue(BlockStateProperties.WATERLOGGED)) {
@@ -82,13 +82,17 @@ public class DisguisableBlockEntity extends CustomizableBlockEntity {
 				level.updateNeighborsAt(worldPosition, state.getBlock());
 			}
 		}
-		else {
+		else
 			ClientHandler.DISGUISED_BLOCK_RENDER_DELEGATE.removeDelegateOf(be);
-			DisguisableBlock.getDisguisedBlockStateFromStack(level, stack).ifPresent(disguisedState -> {
-				if (disguisedState.getLightEmission(level, worldPosition) > 0)
-					level.getChunkSource().getLightEngine().checkBlock(worldPosition);
-			});
-		}
+
+		DisguisableBlock.getDisguisedBlockStateFromStack(level, stack).ifPresent(disguisedState -> {
+			if (disguisedState.getLightEmission(level, worldPosition) > 0) {
+				AuxiliaryLightManager manager = level.getAuxLightManager(worldPosition);
+
+				manager.removeLightAt(worldPosition);
+				manager.setLightAt(worldPosition, state.getLightEmission(level, worldPosition));
+			}
+		});
 	}
 
 	@Override