+++ Changelog.md
@@ -43,6 +43,7 @@
 - Fix: Modules that are inserted in reinforced blocks get deleted when unreinforcing these blocks
 - Fix: Mobs can be converted multiple times by an Electrified Iron Fence (Gate)
 - Fix: The Electrified Iron Fence Gate can be used as fuel in a furnace
+- Fix: Monsters can spawn near blocks disguised as a light-emitting block
 - Removed: Ability to change the subpage in the SecurityCraft Manual using the arrow keys. This is still possible by scrolling while holding down CTRL
 
 --------------------------Changelog for v1.9.8-beta2 of SecurityCraft--------------------------